fit norminette and split code
