+++ src/parse/parse.c
@@ -6,61 +6,13 @@
 /*   By: jaemjeon <jaemjeon@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 18:54:48 by jaemjeon          #+#    #+#             */
-/*   Updated: 2023/05/13 00:43:40 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/05/14 16:19:00 by jaemjeon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lexer.h"
 #include "parse.h"
-
-void	clear_parse_tree(t_parse *root, void (*f)(void *))
-{
-	if (root == NULL)
-		return ;
-	clear_parse_tree(root->left, f);
-	clear_parse_tree(root->right, f);
-	ft_lstclear(&root->node, f);
-	free(root);
-}
-
-void	debug_print_parse_tree(t_parse *parse_tree)
-{
-	if (parse_tree == NULL)
-		return ;
-	printf("debug\n");
-	if (parse_tree->is_subshell)
-		printf("this is SUBSHELL_NODE\n");
-	ft_lstiter(parse_tree->node, debug);
-	debug_print_parse_tree(parse_tree->left);
-	debug_print_parse_tree(parse_tree->right);
-}
-
-t_lex_lst	*get_type_node(t_lex_lst *token_lst, t_e_lex type)
-{
-	int		dquote_flag;
-	int		subshell_depth;
-	t_s_lex	*lex_data;
-
-	dquote_flag = 0;
-	subshell_depth = 0;
-	while (token_lst)
-	{
-		lex_data = token_lst->content;
-		if (lex_data->type == LEXEME_DQUOTE)
-			dquote_flag ^= 1;
-		else if (dquote_flag == 0)
-		{
-			if (lex_data->type == LEXEME_PARENTHESIS_OPEN)
-				subshell_depth++;
-			else if (lex_data->type == LEXEME_PARENTHESIS_CLOSE)
-				subshell_depth--;
-			else if (subshell_depth == 0 && lex_data->type == type)
-				return (token_lst);
-		}
-		token_lst = token_lst->next;
-	}
-	return (NULL);
-}
+#include "parse_inner_util.h"
 
 t_lex_lst	*get_root_node(t_lex_lst **token_lst)
 {
@@ -76,105 +28,6 @@ t_lex_lst	*get_root_node(t_lex_lst **token_lst)
 	if (root_node)
 		return (root_node);
 	return (NULL);
-}
-
-t_parse	*make_new_node(t_lex_lst *token_lst)
-{
-	t_parse	*root;
-
-	if (token_lst == NULL)
-		return (NULL);
-	root = ft_calloc(1, sizeof(t_parse));
-	if (root == NULL)
-		return (NULL);
-	root->node = token_lst;
-	return (root);
-}
-
-int	has_subshell(t_lex_lst *token_lst)
-{
-	int			dquote_flag;
-	t_s_lex		*lex_data;
-
-	dquote_flag = 0;
-	while (token_lst)
-	{
-		lex_data = token_lst->content;
-		if (lex_data->type == LEXEME_DQUOTE)
-			dquote_flag ^= 1;
-		else if (dquote_flag == 0)
-		{
-			if (lex_data->type == LEXEME_PARENTHESIS_OPEN)
-				return (1);
-		}
-		token_lst = token_lst->next;
-	}
-	return (0);
-}
-
-t_lex_lst	*get_before_first_subshell_open(
-	t_lex_lst *token_lst,
-	t_lex_lst **first_open
-)
-{
-	int		dquote_flag;
-	t_s_lex	*lex_data;
-
-	dquote_flag = 0;
-	lex_data = token_lst->content;
-	if (lex_data->type == LEXEME_PARENTHESIS_OPEN)
-	{
-		*first_open = token_lst;
-		return (NULL);
-	}
-	while (token_lst->next)
-	{
-		lex_data = token_lst->content;
-		if (lex_data->type == LEXEME_DQUOTE)
-			dquote_flag ^= 1;
-		else if (dquote_flag == 0 && lex_data->type == LEXEME_PARENTHESIS_OPEN)
-		{
-			*first_open = token_lst->next;
-			return (token_lst);
-		}
-		token_lst = token_lst->next;
-	}
-	return (NULL);
-}
-
-t_lex_lst	*get_before_last_subshell_close(t_lex_lst *token_lst)
-{
-	int			dquote_flag;
-	t_s_lex		*lex_data;
-	t_lex_lst	*last_before_close;
-
-	dquote_flag = 0;
-	while (token_lst->next)
-	{
-		lex_data = token_lst->next->content;
-		if (lex_data->type == LEXEME_DQUOTE)
-			dquote_flag ^= 1;
-		else if (dquote_flag == 0 && lex_data->type == LEXEME_PARENTHESIS_CLOSE)
-			last_before_close = token_lst;
-		token_lst = token_lst->next;
-	}
-	return (last_before_close);
-}
-
-t_lex_lst	*make_new_token_lst(t_lex_lst *start, t_lex_lst *end)
-{
-	t_lex_lst	*new_token_lst;
-
-	new_token_lst = NULL;
-	if (start == NULL || end == NULL)
-		return (NULL);
-	while (start != end)
-	{
-		ft_lstadd_back(&new_token_lst, ft_lstnew(start->content));
-		start = start->next;
-	}
-	ft_lstadd_back(&new_token_lst, ft_lstnew(start->content));
-	return (new_token_lst);
 }
 
 void	split_by_subshell(
@@ -219,147 +72,9 @@ t_parse	*subshell_parsing(t_lex_lst *token_lst)
 	root->is_subshell = 1;
 	split_by_subshell(token_lst, &root->left, &root->right, &root->node);
 	prev_root = root->left;
-	// printf("BEFORE SUBSHELL INNER PARSING!!\n");
-	// ft_lstiter(root->left->node, debug);
-	// printf("\n\n");
 	root->left = parse(root->left->node);
 	free(prev_root);
 	return (root);
-}
-
-t_lex_lst	*get_left_start(t_lex_lst *start, t_lex_lst *root)
-{
-	t_lex_lst	*left_start;
-
-	left_start = start;
-	while (start->next)
-	{
-		if (start->next == root)
-		{
-			start->next = NULL;
-			return (left_start);
-		}
-		start = start->next;
-	}
-	return (left_start);
-}
-
-t_lex_lst	*get_right_start(t_lex_lst *root)
-{
-	t_lex_lst	*right_start;
-
-	right_start = root->next;
-	root->next = NULL;
-	return (right_start);
-}
-
-static void	loop_check_match(
-	t_list *token,
-	int *dquote_flag,
-	int *parenthesis_dep
-)
-{
-	t_s_lex	*cur_token_data;
-
-	while (token)
-	{
-		cur_token_data = token->content;
-		if (cur_token_data->type == LEXEME_DQUOTE)
-			*dquote_flag ^= 1;
-		else if ((cur_token_data->type == LEXEME_PARENTHESIS_OPEN || \
-				cur_token_data->type == LEXEME_PARENTHESIS_CLOSE) && \
-				*dquote_flag == 0)
-		{
-			if (cur_token_data->type == LEXEME_PARENTHESIS_OPEN)
-				(*parenthesis_dep)++;
-			else
-				(*parenthesis_dep)--;
-		}
-		token = token->next;
-	}
-}
-
-int	check_quote_parenthesis_match_error(t_list *token)
-{
-	int	dquote_flag;
-	int	parenthesis_dep;
-
-	dquote_flag = 0;
-	parenthesis_dep = 0;
-	loop_check_match(token, &dquote_flag, &parenthesis_dep);
-	if (dquote_flag == 0 && parenthesis_dep == 0)
-		return (0);
-	else
-		return (1);
-}
-
-int	check_empty_subshell(t_lex_lst *lexlst)
-{
-	t_s_lex	*front_lex_data;
-	t_s_lex	*rear_lex_data;
-
-	while (lexlst && lexlst->next)
-	{
-		rear_lex_data = lexlst->content;
-		front_lex_data = lexlst->next->content;
-		if (rear_lex_data->type == LEXEME_PARENTHESIS_OPEN
-			&& front_lex_data->type == LEXEME_PARENTHESIS_CLOSE)
-			return (1);
-		lexlst = lexlst->next;
-	}
-	return (0);
-}
-
-static int	is_about_string_token(t_lex_lst *token)
-{
-	const t_s_lex	*lex_data = token->content;
-
-	return (lex_data->type == LEXEME_STRING
-		|| lex_data->type == LEXEME_ENVIRONMENT
-		|| lex_data->type == LEXEME_WILDCARD
-		|| lex_data->type == LEXEME_DQUOTE);
-}
-
-static int	is_ifs_token(t_lex_lst *token)
-{
-	const t_s_lex	*lex_data = token->content;
-
-	return (lex_data->type == LEXEME_IFS);
-}
-
-void	delete_useless_ifs(t_lex_lst *rear)
-{
-	t_lex_lst	*front;
-	t_lex_lst	*mid;
-	t_lex_lst	*to_delete;
-
-	if (ft_lstsize(rear) < 3)
-		return ;
-	mid = rear->next;
-	front = mid->next;
-	while (front)
-	{
-		if (!is_about_string_token(mid) && is_ifs_token(front))
-		{
-			mid->next = front->next;
-			to_delete = front;
-			front = front->next;
-			ft_lstdelone(to_delete, free);
-			continue ;
-		}
-		else if (is_ifs_token(mid) && !is_about_string_token(front))
-		{
-			to_delete = mid;
-			rear->next = front;
-			ft_lstdelone(to_delete, free);
-			mid = front;
-			front = front->next;
-			continue ;
-		}
-		front = front->next;
-		mid = mid->next;
-		rear = rear->next;
-	}
 }
 
 t_parse	*parse(t_lex_lst *token_lst)
